Refactor config imports and handling to be inside the config module
